Allow adding custom Value objects to the Data class
And made the default Value classes public. And updated the README a bit.

+++ include/empty.h
@@ -62,7 +62,7 @@ public:
      * 
      *  @return int
      */
-    virtual int size() override
+    virtual size_t size() override
     {
         return 0;
     }
